requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-appconf==1.0.4
django-bootstrap4==3.0.1
django-cleanup==5.2.0
django-compressor==2.4.1
django-environ==0.4.5
django-pandas==0.6.4
django-sass==1.0.1
django-sass-processor==1.0.1
djangorestframework==3.12.4
environ==1.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==2.1.1
libsass==0.21.0
numpy==1.20.3
oauthlib==3.1.1
pandas==1.2.4
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
rcssmin==1.0.6
requests==2.25.1
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
